Resolvendo problemas e adicionando um front simples

+++ bancoforyou/src/main/java/org/foryou/bancoforyou/User/UserMinDTO.java
@@ -40,7 +40,7 @@ public class UserMinDTO{
         return idade;
     }
 
-    public Double getSaldo(){
+    public BigDecimal getSaldo(){
         return saldo;
     }
